moved user related functions to UserController class
