+++ src/main/java/com/codecool/stockhub/controller/CompanyController.java
@@ -1,13 +1,9 @@
 package com.codecool.stockhub.controller;
 
 
-import com.codecool.stockhub.model.User;
 import com.codecool.stockhub.service.HTTPConnection;
-import com.codecool.stockhub.service.UserList;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletResponse;
@@ -19,8 +15,7 @@ public class CompanyController {
     @Autowired
     private HTTPConnection httpConnection;
 
-    @Autowired
-    private UserList userList;
+
 
 
     @GetMapping("/companies")
@@ -30,11 +25,7 @@ public class CompanyController {
         return httpConnection.getContent();
     }
 
-    @PostMapping(value = "/add")
-    public User addUser(@RequestBody User user){
-        userList.registerUser(user);
-        return user;
-    }
+
